Refactor: let VoteHandler deal with all vote changes
